Update streamlit_txt_to_excel_app.py
Converted 'Subcuenta' into integer to be able to filter the rows to be dropped

+++ streamlit_txt_to_excel_app.py
@@ -56,7 +56,8 @@ if st.button("Convert"):
             df.loc[row, 'Compañía'] = df.loc[row, 'Cuenta_Total'][:4]
             df.loc[row, 'Num Centro'] = df.loc[row, 'Cuenta_Total'][5:12]
             df.loc[row, 'Cuenta'] = df.loc[row, 'Cuenta_Total'][13:17]
-            df.loc[row, 'Subcuenta'] = df.loc[row, 'Cuenta_Total'][18:24]
+            # Extract and convert to integer
+            df.loc[row, 'Subcuenta'] = df.loc[row, 'Cuenta_Total'][18:24].astype(int)
             
         # Original df length
         original_length = len(df)
